Server/Application/requirements.txt
pdfminer.six==20221105
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.0
pypdf==3.17.4  